fix: Validator 삭제 및 필요한 클래스의 메소드로 변경
