+++ src/main/kotlin/calculator/Calculator.kt
@@ -3,7 +3,7 @@ package calculator
 object Calculator {
     fun execute(input: String?): String {
         return try {
-            Validator.inputValidate(input)
+            inputValidate(input)
             val result = InputStringParser(input!!).getSeperatedValue().sum()
             result.toString()
         } catch (e: RuntimeException) {
@@ -13,4 +13,20 @@ object Calculator {
             }
         }
     }
+
+    private fun inputValidate(input: String?) {
+        validateInputIsNullOrBlank(input)
+        validateInputIsOnlyNumber(input!!)
+    }
+
+    private fun validateInputIsNullOrBlank(input: String?) {
+        if (input.isNullOrEmpty()) throw InvalidInputException(ZERO)
+        if (input.isBlank()) throw InvalidInputException(ZERO)
+    }
+
+    private fun validateInputIsOnlyNumber(input: String) {
+        if (input.filterNot { it.isDigit() }.isEmpty()) throw InvalidInputException(input)
+    }
+
+    private const val ZERO = "0"
 }
